Move fee display component to ui-signer
