--- packages/app-transfer/src/types.ts
@@ -1,15 +1,0 @@
-// Copyright 2017-2019 @polkadot/app-toolbox authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-import BN from 'bn.js';
-
-export type Fees = {
-  hasAvailable: boolean,
-  isCreation: boolean,
-  isNoEffect: boolean,
-  isRemovable: boolean,
-  isReserved: boolean,
-  txfees: BN,
-  txtotal: BN
-};
